walletdb sigma mint overflow bug

+++ src/wallet/sigmamint.h
@@ -18,7 +18,7 @@ private:
     uint256 txid;
     int nHeight;
     int nId;
-    int denom;
+    int64_t denom;
     bool isUsed;
 
 public:
@@ -30,7 +30,7 @@ public:
         sigma::IntegerToDenomination(denom, value);
         return value;
     }
-    int GetDenominationValue() const {
+    int64_t GetDenominationValue() const {
         return denom;
     }
     uint32_t GetCount() const { return nCount; }
@@ -47,7 +47,7 @@ public:
         sigma::DenominationToInteger(value, denom);
         this->denom = denom;
     };
-    void SetDenominationValue(const int& denom) { this->denom = denom; }
+    void SetDenominationValue(const int64_t& denom) { this->denom = denom; }
     void SetHeight(const int& nHeight) { this->nHeight = nHeight; }
     void SetId(const int& nId) { this->nId = nId; }
     void SetNull();
